Issue 6849: LTS - Fix code coverage
